+++ kernel/cobalt/arch/arm64/ipipe/include/asm/xenomai/machine.h
@@ -51,7 +51,7 @@
 
 static inline __attribute_const__ unsigned long ffnz(unsigned long ul)
 {
-	int __r;
+	unsigned long __r;
 
 	/* zero input is not valid */
 	XENO_WARN_ON(COBALT, ul == 0);
